uint int comparison fix in sound.cpp

+++ src/engine/interface/sound.cpp
@@ -869,7 +869,7 @@ int playsound(int n, const vec *loc = nullptr, extentity *ent = nullptr, int fla
             return -1;
         }
     }
-    if(static_cast<int>(channels.size()) > chanid)
+    if(channels.size() > static_cast<size_t>(chanid))
     {
         SoundChannel &chan = channels[chanid];
         if(sounds.playing(chan, config))
